Bug 1974334: Change SMILAnimationFunction::CompareTo to return int32_t instead of int8_t. r=longsonr
With this change to CompareTo, its return-type now matches that of
nsContentUtils::CompareTreePosition, which makes sense because we directly call
that function (and return the result) in CompareTo's final statement.

Previously, that call could trigger integer overflow (and the wrong comparison
outcome as a result).

Differential Revision: https://phabricator.services.mozilla.com/D255606

UltraBlame original commit: 77a6714a4cca8c60b3d19e362acb976620e082d5

+++ dom/smil/SMILAnimationFunction.h
@@ -140,8 +140,9 @@ class SMILAnimationFunction {
 
 
 
-  int8_t CompareTo(const SMILAnimationFunction* aOther,
-                   nsContentUtils::NodeIndexCache& aCache) const;
+
+  int32_t CompareTo(const SMILAnimationFunction* aOther,
+                    nsContentUtils::NodeIndexCache& aCache) const;
